Major re-work of the CRC model.
Turns out that CRCing the high-level message doesn't work because a
single bit error changes 2…16 destination bits, thus a CRCn ends up as
a random-number test and has a 1/2^n false positive rate (actually
somewhat less because many errors result in an illegal transition)
which is not acceptable for a protocol that's supposed to control
real-world hardware.

Thus we now CRC-check the actual bits on the wire, transmit the CRC
after the end marker instead of before it, and compare calculated with
incoming CRC instead of running the CRC through the calculator and
ending up with zero.

We also dump the CRC6 because it doesn't make as much sense: We need to
run more bits through the calculator, which causes the block to be too
small to work out well. Also, the CRC needs to be in a frame of its own
anyway, in order to prevent convoluted coding and to allow generating a
message on the fly. This also means that we don't need to calculate two
CRCs in parallel, and we can use a CRC table tailored to the number of
wires. Both cut down on code space.

To compensate, an end frame that has >7 bits of filler now carries an
implicit end marker: fortunately, there's enough space for the remaining
bits without having to stretch the end-of-message marker. Thus we no
longer need an explicit the "last byte is valid" flag bit and can save
3…1 end-of-message markers.