PWM threshold for adding control to buffer PID
The idea is that above this threshold, more water returns from the heat
pump than goes to the radiators. Thus if the date is below this threshold,
possibly-cold water from the buffer is mixed in.

In theory, this depends on the setting of the radiator mixer, which we don't
know. However, the assumption is that this is relevant during ramp-up
where the mixer is set to 100% anyway.