lib.cmd: Return stream object
The result of `moat.lib.cmd.CmdHandler`'s `cmd` call is now a Stream
object.  This object has gained a "normal" iterator which returns
the data items *if* the result contains no keywords.

The effect is that you can still write a command handler that returns
with a value, and a `cmd` call that simply assigns its result to
something, as long as you add a comma to the something:

    val, = await conn.cmd("f.bar",42)

which is equivalent to

    res = await conn.cmd("f.bar",42)
	assert not res.kw
	assert len(res.args) == 1
	val = res.args[0]  # or res[0]

Any other result can be